♻️ refactor: completeStep 함수 분리 #434
- 테스트 헬퍼 함수들을 (completeStep1~5, fillAddress) 공통 헬퍼 파일로 분리
- completeStep5 내 유효성 검사 코드 제거
